proj/requirements.txt
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
numpy>=1.26.0
opencv-python>=4.8.0
glfw>=2.5.9
Pillow>=10.0.0 